# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for task tracker",
        "title": "Task Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/task": {
            "post": {
                "description": "Creates a new task with optional assignee and reviewer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create new task",
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TaskCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created task ID",
                        "schema": {
                            "$ref": "#/definitions/model.TaskCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/task/list": {
            "get": {
                "description": "Return all tasks in system with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "backlog",
                            "in_progress",
                            "review",
                            "done"
                        ],
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "critical"
                        ],
                        "type": "string",
                        "description": "priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "project_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "creator",
                        "name": "creator_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "assignee",
                        "name": "assigned_ids",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "reviewer",
                        "name": "reviewer_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TaskListResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid parameters)",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/task/{id}": {
            "get": {
                "description": "Returns information about a specific task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found",
                        "schema": {
                            "$ref": "#/definitions/model.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid task ID",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates information about the existing task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Task update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data for update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TaskUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete task by ID",
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid task ID",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "description": "Return all users in system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "Users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserDB"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiresponse.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.ProjectDB": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.TaskCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "project_id"
            ],
            "properties": {
                "assigned_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                },
                "deadline": {
                    "type": "string",
                    "example": "2025-07-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Task description"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Task name"
                },
                "priority": {
                    "enum": [
                        "low",
                        "medium",
                        "high",
                        "critical"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskPriority"
                        }
                    ],
                    "example": "medium"
                },
                "project_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                },
                "reviewer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                },
                "status": {
                    "enum": [
                        "backlog",
                        "in_progress",
                        "review",
                        "done"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskStatus"
                        }
                    ],
                    "example": "backlog"
                }
            }
        },
        "model.TaskCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "model.TaskListResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Some description"
                },
                "id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                },
                "name": {
                    "type": "string",
                    "example": "Update user table"
                },
                "priority": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskPriority"
                        }
                    ],
                    "example": "medium"
                },
                "project_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                },
                "project_name": {
                    "type": "string"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskStatus"
                        }
                    ],
                    "example": "backlog"
                }
            }
        },
        "model.TaskPriority": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high",
                "critical"
            ],
            "x-enum-varnames": [
                "TaskPriorityLow",
                "TaskPriorityMedium",
                "TaskPriorityHigh",
                "TaskPriorityCritical"
            ]
        },
        "model.TaskResponse": {
            "type": "object",
            "properties": {
                "assignees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDB"
                    }
                },
                "completed": {
                    "type": "string",
                    "example": "2025-07-01T00:00:00Z"
                },
                "creator": {
                    "$ref": "#/definitions/model.UserDB"
                },
                "deadline": {
                    "type": "string",
                    "example": "2025-07-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Some description"
                },
                "id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                },
                "name": {
                    "type": "string",
                    "example": "Update user table"
                },
                "priority": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskPriority"
                        }
                    ],
                    "example": "medium"
                },
                "project": {
                    "$ref": "#/definitions/model.ProjectDB"
                },
                "reviewers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDB"
                    }
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskStatus"
                        }
                    ],
                    "example": "backlog"
                }
            }
        },
        "model.TaskStatus": {
            "type": "string",
            "enum": [
                "backlog",
                "in_progress",
                "review",
                "done"
            ],
            "x-enum-varnames": [
                "TaskStatusBacklog",
                "TaskStatusInProgress",
                "TaskStatusReview",
                "TaskStatusDone"
            ]
        },
        "model.TaskUpdateRequest": {
            "type": "object",
            "required": [
                "name",
                "project_id"
            ],
            "properties": {
                "assigned_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                },
                "completed_at": {
                    "type": "string",
                    "example": "2025-07-01T00:00:00Z"
                },
                "deadline": {
                    "type": "string",
                    "example": "2025-07-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Some description"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Update user table"
                },
                "priority": {
                    "enum": [
                        "low",
                        "medium",
                        "high",
                        "critical"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskPriority"
                        }
                    ],
                    "example": "medium"
                },
                "project_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                },
                "reviewer_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                    ]
                },
                "status": {
                    "enum": [
                        "backlog",
                        "in_progress",
                        "review",
                        "done"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TaskStatus"
                        }
                    ],
                    "example": "backlog"
                }
            }
        },
        "model.UserDB": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Keycloak": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8155/realms/fmgo/protocol/openid-connect/token"
        }
    },
    "security": [
        {
            "OAuth2Keycloak": []
        }
    ]
}